+++ typescript.js
@@ -1,0 +1,16 @@
+module.exports = {
+
+    overrides: [
+        {
+            files: [
+                "*.ts",
+                "*.tsx",
+            ],
+            extends: [
+                "plugin:@typescript-eslint/recommended",
+                "plugin:import/typescript",
+            ],
+        },
+    ],
+
+};