+++ CHANGELOG.md
@@ -1,8 +1,13 @@
+# 3.0.0 / 2021-06-23
+
+- Add `@typescript-eslint` peerDependencies
+- Add typescript extension `@bimproject/eslint-config/typescript`
+
 # 2.0.0 / 2021-06-07
 
 - Update dependencies
 - Move plugins to peer dependencies
-- Add `eslint-plugin-import` and `eslint-plugin-node` as extend bases
+- Add `eslint-plugin-import` and `eslint-plugin-node` as extensions
 - Set `global-require` to "warn"
 - Add React config `@bimproject/eslint-config/react`